+++ src/main/java/com/sluv/server/domain/user/service/UserService.java
@@ -1,16 +1,6 @@
 package com.sluv.server.domain.user.service;
 
 import com.sluv.server.domain.auth.exception.UserBlockedException;
-import com.sluv.server.domain.celeb.dto.InterestedCelebCategoryResDto;
-import com.sluv.server.domain.celeb.dto.InterestedCelebParentResDto;
-import com.sluv.server.domain.celeb.dto.InterestedCelebPostReqDto;
-import com.sluv.server.domain.celeb.entity.Celeb;
-import com.sluv.server.domain.celeb.entity.CelebCategory;
-import com.sluv.server.domain.celeb.entity.InterestedCeleb;
-import com.sluv.server.domain.celeb.exception.CelebNotFoundException;
-import com.sluv.server.domain.celeb.repository.CelebCategoryRepository;
-import com.sluv.server.domain.celeb.repository.CelebRepository;
-import com.sluv.server.domain.celeb.repository.InterestedCelebRepository;
 import com.sluv.server.domain.closet.dto.ClosetResDto;
 import com.sluv.server.domain.closet.entity.Closet;
 import com.sluv.server.domain.closet.repository.ClosetRepository;
@@ -19,49 +9,25 @@ import com.sluv.server.domain.comment.entity.Comment;
 import com.sluv.server.domain.comment.repository.CommentRepository;
 import com.sluv.server.domain.item.dto.ItemSimpleResDto;
 import com.sluv.server.domain.item.entity.Item;
-import com.sluv.server.domain.item.entity.ItemImg;
-import com.sluv.server.domain.item.entity.RecentItem;
+import com.sluv.server.domain.item.mapper.ItemDtoMapper;
 import com.sluv.server.domain.item.repository.ItemImgRepository;
-import com.sluv.server.domain.item.repository.ItemLikeRepository;
 import com.sluv.server.domain.item.repository.ItemRepository;
 import com.sluv.server.domain.item.repository.ItemScrapRepository;
-import com.sluv.server.domain.item.repository.RecentItemRepository;
-import com.sluv.server.domain.question.dto.QuestionImgSimpleResDto;
 import com.sluv.server.domain.question.dto.QuestionSimpleResDto;
 import com.sluv.server.domain.question.entity.Question;
-import com.sluv.server.domain.question.entity.QuestionBuy;
-import com.sluv.server.domain.question.entity.QuestionFind;
-import com.sluv.server.domain.question.entity.QuestionHowabout;
-import com.sluv.server.domain.question.entity.QuestionRecommend;
-import com.sluv.server.domain.question.entity.QuestionRecommendCategory;
-import com.sluv.server.domain.question.entity.RecentQuestion;
-import com.sluv.server.domain.question.exception.QuestionNotFoundException;
-import com.sluv.server.domain.question.exception.QuestionTypeNotFoundException;
-import com.sluv.server.domain.question.repository.QuestionImgRepository;
-import com.sluv.server.domain.question.repository.QuestionItemRepository;
-import com.sluv.server.domain.question.repository.QuestionLikeRepository;
-import com.sluv.server.domain.question.repository.QuestionRecommendCategoryRepository;
+import com.sluv.server.domain.question.mapper.QuestionDtoMapper;
 import com.sluv.server.domain.question.repository.QuestionRepository;
-import com.sluv.server.domain.question.repository.RecentQuestionRepository;
-import com.sluv.server.domain.user.dto.UserIdDto;
 import com.sluv.server.domain.user.dto.UserMypageResDto;
 import com.sluv.server.domain.user.dto.UserProfileImgReqDto;
-import com.sluv.server.domain.user.dto.UserProfileReqDto;
 import com.sluv.server.domain.user.dto.UserSearchInfoDto;
-import com.sluv.server.domain.user.entity.Follow;
 import com.sluv.server.domain.user.entity.User;
 import com.sluv.server.domain.user.enums.UserStatus;
-import com.sluv.server.domain.user.exception.UserNicknameDuplicatedException;
 import com.sluv.server.domain.user.exception.UserNotFoundException;
 import com.sluv.server.domain.user.repository.FollowRepository;
 import com.sluv.server.domain.user.repository.UserRepository;
 import com.sluv.server.global.common.response.PaginationCountResDto;
 import com.sluv.server.global.common.response.PaginationResDto;
-import com.sluv.server.global.jwt.JwtProvider;
-import jakarta.servlet.http.HttpServletRequest;
-import java.util.Comparator;
 import java.util.List;
-import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -76,117 +42,16 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class UserService {
     private final UserRepository userRepository;
-    private final CelebRepository celebRepository;
     private final FollowRepository followRepository;
-    private final InterestedCelebRepository interestedCelebRepository;
     private final ItemRepository itemRepository;
     private final QuestionRepository questionRepository;
     private final CommentRepository commentRepository;
     private final ItemImgRepository itemImgRepository;
     private final ItemScrapRepository itemScrapRepository;
-    private final RecentItemRepository recentItemRepository;
     private final ClosetRepository closetRepository;
-    private final CelebCategoryRepository celebCategoryRepository;
-    private final ItemLikeRepository itemLikeRepository;
+    private final ItemDtoMapper itemDtoMapper;
+    private final QuestionDtoMapper questionDtoMapper;
 
-    private final RecentQuestionRepository recentQuestionRepository;
-    private final QuestionImgRepository questionImgRepository;
-    private final QuestionItemRepository questionItemRepository;
-    private final QuestionRecommendCategoryRepository questionRecommendCategoryRepository;
-    private final QuestionLikeRepository questionLikeRepository;
-
-    private final JwtProvider jwtProvider;
-
-
-    @Transactional(readOnly = true)
-    public UserIdDto getUserIdByToken(HttpServletRequest request) {
-        String token = jwtProvider.resolveToken(request);
-
-        return UserIdDto.of(jwtProvider.getUserId(token));
-    }
-
-    /**
-     * User가 선택한 관심 셀럽을 검색 관심 샐럼의 상위 카테고리를 기준으로 묶어서 Response
-     */
-    @Transactional(readOnly = true)
-    public List<InterestedCelebCategoryResDto> getInterestedCelebByCategory(User user) {
-        List<Celeb> interestedCelebList = celebRepository.findInterestedCeleb(user);
-
-        List<CelebCategory> categoryList = celebCategoryRepository.findAllByParentIdIsNull();
-        changeCategoryOrder(categoryList);
-
-        return categoryList.stream()
-                // 카테고리별 InterestedCelebCategoryResDto 생성
-                .map(category -> {
-                    List<Celeb> categoryFilterCeleb = getCategoryFilterCeleb(interestedCelebList, category);
-                    return InterestedCelebCategoryResDto.of(category,
-                            convertInterestedCelebParentResDto(categoryFilterCeleb));
-                }).toList();
-    }
-
-    public void postUserFollow(User user, Long userId) {
-        // target이 될 유저 검색
-        User targetUser = userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
-
-        // Follow 여부 확인.
-        Boolean followStatus = followRepository.getFollowStatus(user, targetUser);
-
-        if (followStatus) {
-            followRepository.deleteFollow(user, targetUser);
-        } else {
-            // Follow 정보 등록.
-            followRepository.save(Follow.toEntity(user, targetUser));
-        }
-
-    }
-
-    /**
-     * 관심셀럽 목록에서 category와 일치하는 Celeb을 분류
-     */
-    @Transactional(readOnly = true)
-    public List<Celeb> getCategoryFilterCeleb(List<Celeb> celebList, CelebCategory category) {
-        return celebList.stream().filter(celeb ->
-                // interestedCeleb의 상위 카테고리 id와 카테고리별 묶을 카테고리의 아이디가 일치하는 것만 filtering
-                Objects.equals(
-                        celeb.getCelebCategory().getParent() != null ? celeb.getCelebCategory().getParent().getId()
-                                : celeb.getCelebCategory().getId(), category.getId())).toList();
-    }
-
-    /**
-     * 관심셀럽 목록에서 category와 일치하는 Celeb을 분류
-     */
-    private List<InterestedCelebParentResDto> convertInterestedCelebParentResDto(List<Celeb> celebList) {
-        return celebList.stream().map(InterestedCelebParentResDto::of).toList();
-    }
-
-    public void postInterestedCeleb(User user, InterestedCelebPostReqDto dto) {
-        // 기존에 있는 해당 유저의 관심셀럽 초기화
-        interestedCelebRepository.deleteAllByUserId(user.getId());
-
-        // 초기화 상태에서 다시 추가.
-        List<InterestedCeleb> interestedCelebList = dto.getCelebIdList().stream()
-                .map(celeb -> InterestedCeleb.toEntity(user,
-                        celebRepository.findById(celeb).orElseThrow(CelebNotFoundException::new))).toList();
-
-        interestedCelebRepository.saveAll(interestedCelebList);
-    }
-
-    public void postUserProfile(User user, UserProfileReqDto dto) {
-        User currentUser = userRepository.findById(user.getId()).orElseThrow(UserNotFoundException::new);
-
-        // 닉네임 중복 검사
-        Boolean nicknameExistsStatus = userRepository.existsByNickname(dto.getNickName());
-        if (nicknameExistsStatus) {
-            throw new UserNicknameDuplicatedException();
-        }
-
-        currentUser.changeNickname(dto.getNickName());
-        currentUser.changeProfileImgUrl(dto.getImgUrl());
-
-        if (currentUser.getUserStatus().equals(UserStatus.PENDING_PROFILE)) {
-            currentUser.changeUserStatus(UserStatus.PENDING_CELEB);
-        }
-    }
 
     @Transactional(readOnly = true)
     public UserMypageResDto getUserMypage(User user, Long userId) {
@@ -226,16 +91,10 @@ public class UserService {
 
         List<ItemSimpleResDto> content = itemPage.stream().map(item -> {
             List<Closet> closetList = closetRepository.findAllByUserId(user.getId());
-            return getItemSimpleResDto(item, closetList);
+            return itemDtoMapper.getItemSimpleResDto(item, closetList);
         }).toList();
 
         return PaginationResDto.of(itemPage, content);
-    }
-
-    @Transactional(readOnly = true)
-    private ItemSimpleResDto getItemSimpleResDto(Item item, List<Closet> closetList) {
-        return ItemSimpleResDto.of(item, itemImgRepository.findMainImg(item.getId()),
-                itemScrapRepository.getItemScrapStatus(item, closetList));
     }
 
     @Transactional(readOnly = true)
@@ -255,114 +114,6 @@ public class UserService {
         return PaginationResDto.of(closetPage, content);
     }
 
-    @Transactional(readOnly = true)
-    public PaginationCountResDto<ItemSimpleResDto> getUserRecentItem(User user, Pageable pageable) {
-
-        Page<RecentItem> recentItemPage = recentItemRepository.getUserAllRecentItem(user, pageable);
-
-        List<Closet> closetList = closetRepository.findAllByUserId(user.getId());
-
-        List<ItemSimpleResDto> content = recentItemPage.stream()
-                .map(recentItem -> getItemSimpleResDto(recentItem.getItem(), closetList)).toList();
-
-        return new PaginationCountResDto<>(recentItemPage.hasNext(), recentItemPage.getNumber(), content,
-                recentItemPage.getTotalElements());
-
-    }
-
-    @Transactional(readOnly = true)
-    public PaginationCountResDto<QuestionSimpleResDto> getUserRecentQuestion(User user, Pageable pageable) {
-
-        Page<RecentQuestion> recentQuestionPage = recentQuestionRepository.getUserAllRecentQuestion(user, pageable);
-
-        List<QuestionSimpleResDto> content = recentQuestionPage.stream().map(recentQuestion -> {
-            Question question = questionRepository.findById(recentQuestion.getQuestion().getId())
-                    .orElseThrow(QuestionNotFoundException::new);
-            List<QuestionImgSimpleResDto> imgList = null;
-            List<QuestionImgSimpleResDto> itemImgList = null;
-            List<String> categoryList = null;
-            String celebName = null;
-
-            if (recentQuestion.getQType().equals("Buy")) {
-                // 이미지 Dto 생성
-                imgList = questionImgRepository.findAllByQuestionId(question.getId()).stream()
-                        .map(QuestionImgSimpleResDto::of).toList();
-                // 아이템 이미지 Dto 생성
-                itemImgList = questionItemRepository.findAllByQuestionId(question.getId()).stream()
-                        .map(questionItem -> {
-                            ItemImg mainImg = itemImgRepository.findMainImg(questionItem.getItem().getId());
-                            return QuestionImgSimpleResDto.of(mainImg);
-                        }).toList();
-
-            } else if (recentQuestion.getQType().equals("How")) {
-
-            } else if (recentQuestion.getQType().equals("Recommend")) {
-                // Question 카테고리
-                categoryList = questionRecommendCategoryRepository.findAllByQuestionId(question.getId()).stream()
-                        .map(QuestionRecommendCategory::getName).toList();
-            } else if (recentQuestion.getQType().equals("Find")) {
-                QuestionFind questionFind = (QuestionFind) question;
-                celebName = questionFind.getCeleb() != null ? questionFind.getCeleb().getParent() != null ?
-                        questionFind.getCeleb().getParent().getCelebNameKr() + " " + questionFind.getCeleb()
-                                .getCelebNameKr() : questionFind.getCeleb().getCelebNameKr()
-                        : questionFind.getNewCeleb().getCelebName();
-            } else {
-                throw new QuestionTypeNotFoundException();
-            }
-
-            // Question 좋아요 수
-            Long likeNum = questionLikeRepository.countByQuestionId(question.getId());
-
-            // Question 댓글 수
-            Long commentNum = commentRepository.countByQuestionId(question.getId());
-
-            return QuestionSimpleResDto.of(question, likeNum, commentNum, imgList, itemImgList, categoryList);
-
-
-        }).toList();
-
-        return new PaginationCountResDto<>(recentQuestionPage.hasNext(), recentQuestionPage.getNumber(), content,
-                recentQuestionPage.getTotalElements());
-    }
-
-    @Transactional(readOnly = true)
-    public PaginationCountResDto<ItemSimpleResDto> getUserLikeItem(User user, Pageable pageable) {
-        Page<Item> itemPage = itemRepository.getAllByUserLikeItem(user, pageable);
-
-        List<Closet> closetList = closetRepository.findAllByUserId(user.getId());
-
-        List<ItemSimpleResDto> content = itemPage.stream().map(item -> getItemSimpleResDto(item, closetList)).toList();
-
-        return new PaginationCountResDto<>(itemPage.hasNext(), itemPage.getNumber(), content,
-                itemPage.getTotalElements());
-    }
-
-    @Transactional(readOnly = true)
-    public PaginationResDto<UserSearchInfoDto> getUserFollower(User user, Long userId, Pageable pageable) {
-        // Follower 들 조회
-        Page<User> followerPage = userRepository.getAllFollower(userId, pageable);
-
-        // UserSearchInfoDto로 가공
-        List<UserSearchInfoDto> content = followerPage.stream()
-                .map(follower -> UserSearchInfoDto.of(follower, followRepository.getFollowStatus(user, follower)))
-                .toList();
-
-        return PaginationResDto.of(followerPage, content);
-    }
-
-    @Transactional(readOnly = true)
-    public PaginationResDto<UserSearchInfoDto> getUserFollowing(User user, Long userId, Pageable pageable) {
-        // Following 들 조회
-        Page<User> followerPage = userRepository.getAllFollowing(userId, pageable);
-
-        // UserSearchInfoDto로 가공
-        List<UserSearchInfoDto> content = followerPage.stream()
-                .map(follower -> UserSearchInfoDto.of(follower, followRepository.getFollowStatus(user, follower)))
-                .toList();
-
-        return PaginationResDto.of(followerPage, content);
-    }
-
     public void patchUserProfileImg(User user, UserProfileImgReqDto dto) {
         log.info("User Profile Img Change. User: {}", user.getId());
         log.info("Change URL: {}", dto.getImgUrl());
@@ -377,83 +128,6 @@ public class UserService {
     }
 
     /**
-     * 유저가 좋아요한 Question 게시글 조회
-     */
-    @Transactional(readOnly = true)
-    public PaginationCountResDto<QuestionSimpleResDto> getUserLikeQuestion(User user, Pageable pageable) {
-        Page<Question> questionPage = questionRepository.getUserLikeQuestion(user, pageable);
-
-        List<QuestionSimpleResDto> content = questionPage.stream().map(this::dtoBuildByQuestionType).toList();
-
-        return new PaginationCountResDto<>(questionPage.hasNext(), questionPage.getNumber(), content,
-                questionPage.getTotalElements());
-    }
-
-    private QuestionSimpleResDto dtoBuildByQuestionType(Question question) {
-        String qType = null;
-        List<QuestionImgSimpleResDto> imgList = null;
-        List<QuestionImgSimpleResDto> itemImgList = null;
-        String celebName = null;
-        List<String> categoryList = null;
-
-        if (question instanceof QuestionBuy) { // 1. question이 QuestionBuy 일 경우
-            // 이미지 DTO 생성
-            List<QuestionImgSimpleResDto> imgSimpleList = questionImgRepository.findAllByQuestionId(question.getId())
-                    .stream().map(QuestionImgSimpleResDto::of).toList();
-            // 아이템 이미지 DTO 생성
-            List<QuestionImgSimpleResDto> itemImgSimpleList = questionItemRepository.findAllByQuestionId(
-                    question.getId()).stream().map(questionItem -> {
-                ItemImg mainImg = itemImgRepository.findMainImg(questionItem.getItem().getId());
-                return QuestionImgSimpleResDto.of(mainImg);
-            }).toList();
-
-            qType = "Buy";
-            imgList = imgSimpleList;
-            itemImgList = itemImgSimpleList;
-
-        } else if (question instanceof QuestionFind questionFind) { // 2. question이 QuestionFind 일 경우
-            qType = "Find";
-            celebName = questionFind.getCeleb() != null ? questionFind.getCeleb().getParent() != null ?
-                    questionFind.getCeleb().getParent().getCelebNameKr() + " " + questionFind.getCeleb()
-                            .getCelebNameKr() : questionFind.getCeleb().getCelebNameKr()
-                    : questionFind.getNewCeleb().getCelebName();
-
-        } else if (question instanceof QuestionHowabout) { // 3. question이 QuestionHowabout 일 경우
-            qType = "How";
-
-        } else if (question instanceof QuestionRecommend) { // 4. question이 QuestionRecommend 일 경우
-            List<String> categoryNameList = questionRecommendCategoryRepository.findAllByQuestionId(question.getId())
-                    .stream().map(QuestionRecommendCategory::getName).toList();
-            qType = "Recommend";
-            categoryList = categoryNameList;
-
-        } else {
-            throw new QuestionTypeNotFoundException();
-        }
-
-        // Question 좋아요 수
-        Long likeNum = questionLikeRepository.countByQuestionId(question.getId());
-
-        // Question 댓글 수
-        Long commentNum = commentRepository.countByQuestionId(question.getId());
-
-        return QuestionSimpleResDto.of(question, likeNum, commentNum, imgList, itemImgList, categoryList);
-    }
-
-    /**
-     * 유저가 좋아요한 댓글 목록 조회
-     */
-    @Transactional(readOnly = true)
-    public PaginationCountResDto<CommentSimpleResDto> getUserLikeComment(User user, Pageable pageable) {
-        Page<Comment> commentPage = commentRepository.getUserAllLikeComment(user, pageable);
-
-        List<CommentSimpleResDto> content = commentPage.stream().map(CommentSimpleResDto::of).toList();
-
-        return new PaginationCountResDto<>(commentPage.hasNext(), commentPage.getNumber(), content,
-                commentPage.getTotalElements());
-    }
-
-    /**
      * 현재 유저가 업로드한 아이템 조회
      */
     @Transactional(readOnly = true)
@@ -463,7 +137,8 @@ public class UserService {
         // 현재 유저의 옷장 검색
         List<Closet> closetList = closetRepository.findAllByUserId(user.getId());
         // content 제작
-        List<ItemSimpleResDto> content = itemPage.stream().map(item -> getItemSimpleResDto(item, closetList)).toList();
+        List<ItemSimpleResDto> content = itemPage.stream()
+                .map(item -> itemDtoMapper.getItemSimpleResDto(item, closetList)).toList();
 
         return new PaginationCountResDto<>(itemPage.hasNext(), itemPage.getNumber(), content,
                 itemPage.getTotalElements());
@@ -476,7 +151,9 @@ public class UserService {
     public PaginationCountResDto<QuestionSimpleResDto> getUserUploadQuestion(User user, Pageable pageable) {
         Page<Question> questionPage = questionRepository.getUserAllQuestion(user, pageable);
 
-        List<QuestionSimpleResDto> content = questionPage.stream().map(this::dtoBuildByQuestionType).toList();
+        List<QuestionSimpleResDto> content = questionPage.stream()
+                .map(questionDtoMapper::dtoBuildByQuestionType)
+                .toList();
 
         return new PaginationCountResDto<>(questionPage.hasNext(), questionPage.getNumber(), content,
                 questionPage.getTotalElements());
@@ -505,60 +182,10 @@ public class UserService {
 
     }
 
-    /**
-     * 특정 유저가 선택한 관심 Celeb을 조회 CelebCategory를 기준으로 그룹핑 카테고리를 기준으로 조회
-     */
-    @Transactional(readOnly = true)
-    public List<InterestedCelebCategoryResDto> getTargetUserInterestedCelebByCategory(Long userId) {
-        User user = userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
-        List<Celeb> interestedCelebList = celebRepository.findInterestedCeleb(user);
-
-        List<CelebCategory> categoryList = celebCategoryRepository.findAllByParentIdIsNull();
-        categoryList.sort(Comparator.comparing(CelebCategory::getName));
-        changeCategoryOrder(categoryList);
-
-        return categoryList.stream()
-                // 카테고리별 InterestedCelebCategoryResDto 생성
-                .map(category -> {
-                    List<Celeb> categoryFilterCeleb = getCategoryFilterCeleb(interestedCelebList, category);
-                    return InterestedCelebCategoryResDto.of(category,
-                            convertInterestedCelebParentResDto(categoryFilterCeleb));
-                }).toList();
-    }
-
-    /**
-     * 가수 -> 배우 -> 방송인 -> 스포츠인 -> 인플루언서 순서로 변
-     */
-
-    private void changeCategoryOrder(List<CelebCategory> categoryList) {
-        categoryList.sort(Comparator.comparing(CelebCategory::getName));
-
-        CelebCategory tempCategory = categoryList.get(1);
-        categoryList.set(1, categoryList.get(2));
-        categoryList.set(2, tempCategory);
-    }
-
     public void checkUserStatue(User user) {
 
         if (user.getUserStatus().equals(UserStatus.BLOCKED)) {
             throw new UserBlockedException();
         }
-    }
-
-    /**
-     * 특정 유저가 선택한 관심 Celeb을 조회 CelebCategory를 기준으로 그룹핑 유저가 등록한 순서대로 조회
-     */
-    @Transactional(readOnly = true)
-    public List<InterestedCelebParentResDto> getTargetUserInterestedCelebByPostTime(Long userId) {
-        User user = userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
-        return celebRepository.findInterestedCeleb(user).stream().map(InterestedCelebParentResDto::of).toList();
-    }
-
-    /**
-     * User가 선택한 관심 셀럽을 검색 관심 셀럽 선택순서를 기준으로 조회
-     */
-    @Transactional(readOnly = true)
-    public List<InterestedCelebParentResDto> getInterestedCelebByPostTime(User user) {
-        return celebRepository.findInterestedCeleb(user).stream().map(InterestedCelebParentResDto::of).toList();
     }
 }
